Split conv_factory.hpp into separate files
